refactor: move screen_buffer.[h,c]pp to ScreenBuffer.[h,c]pp
- Move the files src/Editor/screen_buffer/screen_buffer.* to
  src/ScreenBuffer/ScreenBuffer.*
