+++ src/CMakeLists.txt
@@ -38,8 +38,8 @@ target_sources(kilo
         "${PROJECT_SOURCE_DIR}/src/Terminal/window/window.hpp"
         "${PROJECT_SOURCE_DIR}/src/Terminal/window/window.cpp"
 
-        "${PROJECT_SOURCE_DIR}/src/Editor/screen_buffer/screen_buffer.hpp"
-        "${PROJECT_SOURCE_DIR}/src/Editor/screen_buffer/screen_buffer.cpp"
+        "${PROJECT_SOURCE_DIR}/src/ScreenBuffer/ScreenBuffer.hpp"
+        "${PROJECT_SOURCE_DIR}/src/ScreenBuffer/ScreenBuffer.cpp"
     
         main.cpp
 )
