+++ src/Editor/Editor.cpp
@@ -26,10 +26,10 @@
 #include "Constants/Constants.hpp"
 #include "Cursor/Cursor.hpp"
 #include "Offset/Offset.hpp"
+#include "ScreenBuffer/ScreenBuffer.hpp"
 #include "Terminal/File.hpp"
 #include "Terminal/window/window.hpp"
 #include "Utilities/Utilities.hpp"
-#include "screen_buffer/screen_buffer.hpp"
 #include <algorithm>
 #include <cstdio>
 #include <cstdlib>
